Create 912. Sort an Array
In Quicksort, it is much important that the ending condition of the while loop is left <= right not left < right. Because if the left pointer and right pointer are in the same position, then left is not < right, the while loop will stop here. In this case, the recursion partition could create two partitions sharing the same nums[left/right].

In addition, if partition is [1,2], left on 1, right on 2, pivot = 1. nums[left] not > 1, keep still.  num[right] > 1, move to 1. In this case, left = right, stop moving and quit the loop, new right part of the partition is still [1,2], which will cause maxmium recursion depth exceed which is the same as the stack overflow in Java.